Каталог, категория, товар. Удалена поддеркжа работы с памятью, оставлен только вариант с БД Postgres.
