+++ src/main/java/com/colvir/shop/config/Config.java
@@ -1,9 +1,5 @@
 package com.colvir.shop.config;
 
-import com.colvir.shop.repository.CatalogRepository;
-import com.colvir.shop.repository.CategoryRepository;
-import com.colvir.shop.repository.ProductRepository;
-import com.colvir.shop.repository.impl.*;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -19,7 +15,6 @@ import java.util.Properties;
 @Configuration
 public class Config {
 
-    private final String repositoryMode;
     private final String driver;
     private final String url;
     private final String username;
@@ -28,7 +23,6 @@ public class Config {
     private final String hibernateShowSql;
 
     public Config(
-            @Value("${repository.mode}") String repositoryMode,
             @Value("${database.driver}") String driver,
             @Value("${database.url}") String url,
             @Value("${database.username}") String username,
@@ -36,7 +30,6 @@ public class Config {
             @Value("${hibernate.dialect}") String hibernateDialect,
             @Value("${hibernate.show.sql}") String hibernateShowSql) {
 
-        this.repositoryMode = repositoryMode;
         this.driver = driver;
         this.url = url;
         this.username = username;
@@ -59,39 +52,6 @@ public class Config {
     @Bean
     public JdbcTemplate getJdbcTemplate() {
         return new JdbcTemplate(getDataSource());
-    }
-
-    @Bean
-    public CatalogRepository getCatalogRepository () {
-
-        return switch (repositoryMode) {
-            case "memory" -> new CatalogRepositoryMemoryImpl();
-            case "postgres" -> new CatalogRepositoryPostgresImpl(getJdbcTemplate());
-            default ->
-                    throw new RuntimeException(String.format("Режим репозитория %s не поддерживается.", repositoryMode));
-        };
-    }
-
-    @Bean
-    public CategoryRepository getCategoryRepository () {
-
-        return switch (repositoryMode) {
-            case "memory" -> new CategoryRepositoryMemoryImpl();
-            case "postgres" -> new CategoryRepositoryPostgresImpl(getLocalSessionFactoryBean().getObject());
-            default ->
-                    throw new RuntimeException(String.format("Режим репозитория %s не поддерживается.", repositoryMode));
-        };
-    }
-
-    @Bean
-    public ProductRepository getProductRepository () {
-
-        return switch (repositoryMode) {
-            case "memory" -> new ProductRepositoryMemoryImpl();
-            case "postgres" -> new ProductRepositoryPostgresImpl(getLocalSessionFactoryBean().getObject());
-            default ->
-                    throw new RuntimeException(String.format("Режим репозитория %s не поддерживается.", repositoryMode));
-        };
     }
 
     private Properties getHibernateProperties() {
