+++ src/main/resources/application.properties
@@ -1,7 +1,5 @@
 spring.application.name=shop
 server.servlet.context-path=/api
-#repository.mode=memory
-repository.mode=postgres
 database.driver=org.postgresql.Driver
 database.url=jdbc:postgresql://localhost:5432/postgres
 database.username=postgres
